Added feel on|off and 'new n' to create empty list with 'n' points.
Also changed when SetStatus() is called. It now needs to be called
explicitly. Previously it was called from SetFocus(). This is now mostly
redundant, as any method which changes an instance will also set the
status when finished. There may be knock effects from this which I'll
clean up as I find them.