Scale Visibility, now we're getting somewhere.
On-screen keyboard shows in/out mappings using text based
line drawing characters. One mapping is shown every step,
but there's an option to pause the display at any time. The
idea is to use this to analyse what's going on, setting step
and quantum to make the display readable, but you wouldn't
necessarily find this useful under normal circumstances.