Midi Phasing, usable!
Following on from midi file import to real time lists, phasing involves
setting independent play rates for each list. This entailed adding a
play position counter to each list and changing the Step() function
to work with it. Behaviour of local loop quantum and start has also changed,
but both features still provide the ability to loop a sub-section of the
captured or imported list.

There's also a new command - "real time rate [rate] [step]" - which sets
play rate, or a range of rates, for all real time lists in a pattern.