Midi Phasing, further tweaks
Actually not just for phasing. I fixed a bug in the
Step2() function which meant that loop start (and
things in general) wasn't working for small values
of list quantum.

Also made some attempts to deal with boundary errors
introduced by floating point inconsistencies. Even
with doubles, attempting to perform range matching
and other logic is proving consistently troublesome.
Do I have to do everything with integers and commit,
say, to limiting accuracy to 1000th of a beat.