Three builds succeed, doesn't work on target.
Three or four days hacking out (with conditional defines)
try/catch to build with no-exceptins; stoi()/stod() replaced
with strol()/strtod(), because the former aren't in the AVR
libs; changing unordered_map<> to key on 'int' because the AVR
STL doesn't want to work with enum types; writing an ANSI
console class to avoid ncurses - the hardest part was finding
out how to interpret incoming escape sequences and how to turn
off ctrl-c (you configure the terminal with termios.h, though I
read through pages and pages on managing signal handlers before
it dawned on my that the answer lay elsewhere).

This builds for AVR, but there's absolutely no sign of life when
I load it. More puzzling, the IDE reports code size of 154860. The
hex file is 438K, and the sum of all the .o files runs into
several megabytes. Is this real, or has something wrapped?

I'll commit here and make some major cuts until things start
working.