RT input in place on MC, but there are problems ...
A timeline is in place to aid with Real Time input capture
but it's problematic (and I haven't pinned down behaviour yet).

To assist with debugging, I've added direct ALSA I/O on the
BLUE PC build to aid with testing, but I've only just got this
working and haven't used it to diagnose anything yet.

This commit has two other issues: Rests in step lists on the
MC seem to affect step timeing, and the last build hangs on
the MC after playing just four notes of a simple step list.