WIP, basic callback routing in place
I just pasted the bulk of the note processing code
into StepListGroup::Step() prior to sorting through
and adapting it. It obviously doesn't build, but I
need to make a backup as I'm about to be interrupted.
Just take this code out if you want a working build
at this point.