WIP, in the middle of a different nowhere.
Fixed a bug or two regarding menu list selection after
delete/copy operations. Now deliberately throwing unhandled
exceptions when trying to the current cursor pos from a menu
list. (This happens if we get to the destructor for the selected
item.)

Anyway, still mulling over lots and lots of things, but need
to make a backup whilst upgrading Ubuntu Studio.