WIP, copy constructors, move constructors *and* assignment operators
Beginning to feel there's a coherent strategy for ItemMenu/MenuList.

Copy constructor just handles reallocation when vectors grow.
(I failed to have the move contructor called for this, despite
specifying 'noexcept' for the constructor and everything it calls.)

Move constructor *is* called when items are deleted from within
a vector. (I didn't know about move constructors.)

Use assignment for actual copy. It means creating a default object
first, but I think that would happen anyway. The downside is that
the move constructor contains pretty much the same code as the copy
constructor.

(In an ideal world, I'd be able to use copy ctor for genuine copy,
move ctor for both kinds of vector size changes, and do without
assignment.)

I'm still in the early stages of debugging this, by the way.